Launching works
Note:  there's a bizarre overlap bug betwen flapjax and Jquery.

It is *not* the $-sign thing that's fixed here.  Rather, something to do with
mapping or filtering or forEaching, deep in the bowels of Jquery, is
overwritten by flapjax or vice versa.

On a flapjax page, if you do:

var domThing = $('<div></div>');
domThing.attr({
  foo: 'bar'
});

Jquery will go into an infinite loop.  Fun times.